Got a lot done.
Set up custom types, the three main functions, added ablity to print to the actual screen, and added the s_scene and s_location types. Now we just need to add the other types, add the main loop that prases user directives, and add the ablity to play audio and display images and character sprites; still have quite a bit to do.