Refactored lua api
We are now using 'full userdata' for scenes and we have get and set functions as well as load and unload functions; we have more of an OOP approach now. Locations still use lightuserdata but we will not give metatables to them and they can only be created and deleted, not edited. Not much has been done to the token processing stage yet since I am still trying to get the 'feel' of the lua api to feel good, but so far, I think it does so I will do some more work on token processing in a bit. And of course I do still have to add 'things' and 'items' and I need to add some brief audio support, but after that I will consider this project a huge success.