Fully implemented locations (aside from the  'things' in them)
Finally got the locationmap function to work as intented, set up callbacks for enter, exit, and invalid, and allowed the 'go' directive to use cardinal directions for the locations. Next I should add audio support to take a break from token phrasing and so I can finish up the media portion of the engine. I may add some more media functions later but they will probably be for polish. Once I'm done with the audio I will try to add 'things' and 'items' and the appropriate lua api and token phrasing for them, and after some polish we will be done and we can move on to our first field test, "The Backrooms."