Added userdata validation system and started work on audio
There is now a userdata validation system which uses the s_data structure which has a 'type' and a 'data' pointer, the data is the actual data and the type allows us to know what type it should be, this does add a little confusion and indirection but makes our program much safer for the lua client user because they now get good error messages when they mess up sine the engine tells them what type it expected and what type it wanted. I also started work on the audio system, gave it a metatable and started up mixer but for some reason it wont work on the wav file I have. I will have to get another file to be sure it is my fault and it doesn't help that the internet isn't working here. TODO: Fix audio.